requirements.txt
Flask
flask-cors
numpy
opencv-python
tensorflow
pillow
requests
scikit-learn